ProofBuilder: moved `t`, const not private to lk
