+++ core/src/main/scala/at/logic/gapt/proofs/lk/ProofBuilder.scala
@@ -18,7 +18,7 @@ import at.logic.gapt.proofs.resolution.ResolutionProof
  *
  * @param proofStack
  */
-class ProofBuilder[Proof] private[lk] ( private val proofStack: Seq[Proof] ) {
+class ProofBuilder[Proof]( private val proofStack: Seq[Proof] ) {
 
   /**
    * Pushes a proof onto the stack.
@@ -49,6 +49,19 @@ class ProofBuilder[Proof] private[lk] ( private val proofStack: Seq[Proof] ) {
     case Seq()            => throw new Exception( "Cannot apply binary inference to empty stack." )
     case p +: Seq()       => throw new Exception( "Cannot apply binary inference to stack with only one element." )
     case p2 +: p1 +: rest => new ProofBuilder( inference( p1, p2 ) +: rest )
+  }
+
+  /**
+   * Applies a ternary inference to the top three elements of the proof stack.
+   *
+   * @param inference A function (NDProof, NDProof, NDProof) => NDProof
+   * @return
+   */
+  def t( inference: ( Proof, Proof, Proof ) => Proof ) = proofStack match {
+    case Seq()                  => throw new Exception( "Cannot apply ternary inference to empty stack." )
+    case p +: Seq()             => throw new Exception( "Cannot apply ternary inference to stack with only one element." )
+    case p2 +: p1 +: Seq()      => throw new Exception( "Cannot apply ternary inference to stack with only two elements." )
+    case p3 +: p2 +: p1 +: rest => new ProofBuilder( inference( p1, p2, p3 ) +: rest )
   }
 
   /**
